+++ changelog.md
@@ -1,4 +1,8 @@
-# Version 0.5.1
+# Version 0.6.0
 
-# Suppressions
-- Suppression des `print()` de debug
+# Ajouts
+- Ajout d'un système ajoutant le format du container dans le container
+- Ajout de la méthode `__create_header()` qui crée le header à partir des éléments suivant :
+  - Format du container
+  - Taille de la table
+- Ajout de la méthode `__get_header()` qui récupère le header, ainsi que le format et la taille de la table